Add types for github route
