+++ backend/package.json
@@ -74,6 +74,7 @@
     },
     "devDependencies": {
         "@babel/preset-typescript": "^7.15.0",
+        "@octokit/types": "^7.5.0",
         "@types/archiver": "^5.3.1",
         "@types/bytes": "^3.1.1",
         "@types/cookie": "^0.4.1",
